refactor(terminal): extract sidebar item components
